:lock: Write on db changed
Just card's owner can delete the card from database

+++ lib/pages/home_page/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
@@ -9,8 +10,10 @@ import 'package:rpg_cards/pages/card_page/card_page.dart';
 import '../card_page/components/tile_list.dart';
 
 class HomePage extends StatelessWidget {
+  HomePage(this.currentUser);
   FirebaseFirestore db = FirebaseFirestore.instance;
   final storage = FirebaseStorage.instance;
+  User? currentUser;
 
   @override
   Widget build(BuildContext context) {
@@ -43,16 +46,16 @@ class HomePage extends StatelessWidget {
                       personagens.forEach((element) {
                         lista_personagens.add(
                           PersonagemBean(
-                              nome: element['nome'],
-                              classe: element['classe'],
-                              arma: element['arma'],
-                              ataque: element['ataque'],
-                              image: element['image'],
-                              imageref: element['imageRef'],
-                              id: element.id,
-                              dado: (element['dado'] == null
-                                  ? '0'
-                                  : element['dado'])),
+                            nome: element['nome'],
+                            classe: element['classe'],
+                            arma: element['arma'],
+                            ataque: element['ataque'],
+                            image: element['image'],
+                            imageref: element['imageRef'],
+                            id: element.id,
+                            userUid: element['userUid'],
+                            dado: element['dado'],
+                          ),
                         );
                       });
 
@@ -69,15 +72,39 @@ class HomePage extends StatelessWidget {
                                 children: [
                                   SlidableAction(
                                     onPressed: (value) {
-                                      db
-                                          .collection('characters')
-                                          .doc(personagens[index].id)
-                                          .delete();
-                                      FirebaseStorage.instance
-                                          .ref()
-                                          .child(
-                                              lista_personagens[index].imageref)
-                                          .delete();
+                                      if (currentUser?.uid ==
+                                          'HBGrv9VtszeVrQBM1MNpOGZxco82' || currentUser?.uid == personagens[index]['userUid']) {
+                                        db
+                                            .collection('characters')
+                                            .doc(personagens[index].id)
+                                            .delete();
+                                        FirebaseStorage.instance
+                                            .ref()
+                                            .child(lista_personagens[index]
+                                                .imageref)
+                                            .delete();
+                                      } else {
+                                        showDialog(
+                                          context: context,
+                                          builder: (BuildContext context) {
+                                            // retorna um objeto do tipo Dialog
+                                            return AlertDialog(
+                                              title: new Text("Erro"),
+                                              content: new Text(
+                                                  "Você não tem permissão!"),
+                                              actions: <Widget>[
+                                                // define os botões na base do dialogo
+                                                new TextButton(
+                                                  child: new Text("Fechar"),
+                                                  onPressed: () {
+                                                    Navigator.of(context).pop();
+                                                  },
+                                                ),
+                                              ],
+                                            );
+                                          },
+                                        );
+                                      }
                                     },
                                     backgroundColor: Colors.redAccent,
                                     icon: Icons.delete,
@@ -106,7 +133,7 @@ class HomePage extends StatelessWidget {
                     context,
                     MaterialPageRoute(
                       builder: (context) {
-                        return AddPage();
+                        return AddPage(currentUser: currentUser);
                       },
                     ),
                   );
